requirements.txt
amqp==2.6.1
asgiref==3.2.10
autopep8==1.4.4
billiard==3.6.3.0
celery==4.4.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.14
django-allauth==0.39.1
django-cors-headers==3.4.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-filter==2.3.0
django-mathfilters==1.0.0
djangorestframework==3.11.0
html5lib==1.1
idna==2.8
kombu==4.6.11
Markdown==3.2.2
mysqlclient==2.0.1
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.2
pycodestyle==2.5.0
PyPDF2==1.26.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
reportlab==3.5.46
requests==2.21.0
requests-oauthlib==1.2.0
six==1.15.0
sqlparse==0.2.4
stripe==2.27.0
urllib3==1.24.2
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.4
